Rename state machine to machine.ts
