+++ packages/react-liveness/src/components/FaceLivenessDetector/service/index.ts
@@ -1,3 +1,3 @@
-export * from './machine';
+export * from './machine/machine';
 export * from './types';
 export * from './utils';
