+++ packages/react-liveness/src/components/FaceLivenessDetector/service/machine/__tests__/index.test.ts
@@ -1,7 +1,7 @@
 import { interpret } from 'xstate';
 import { setImmediate } from 'timers';
 
-import { livenessMachine } from '..';
+import { livenessMachine } from '../machine';
 import {
   FaceLivenessDetectorProps,
   FaceMatchState,
